refactor(profiles): improve controllers and model validations
- Use `find_by` for record retrieval to improve error handling
- Switch to symbol array shorthand for `before_action` filters for readability
- Replace `image_url` with `image_key` for consistency
- Add validation for `stars` attribute in Profile model
- Fix route for showing presentation cards

+++ backend/profile-service/app/controllers/institutions_controller.rb
@@ -1,22 +1,26 @@
 class InstitutionsController < ApplicationController
-  before_action :set_institution, only: %i[:show, :update, :destroy]
-  before_action :set_own_institution, only: %i[:show_own_institution]
+  before_action :set_institution, only: %i[show update destroy]
+  before_action :set_own_institution, only: %i[show_own_institution]
 
   # GET institutions
   def index
-    @institutions = Institution.select(:id, :name, :image_url)
+    @institutions = Institution.select(:id, :name, :image_key)
 
     render json: @institutions
   end
 
   # GET institutions/1
   def show
-    render json: @institution
+    if @institution
+      render json: @institution, status: :ok
+    else
+      render json: { error: 'Unable to fetch institution' }, status: :unprocessable_entity
+    end
   end
 
   # GET profiles/me/institutions
   def show_own_institution
-    render json: { id: @institution.id, name: @institution.name, image_url: @institution.image_url }
+    render json: { id: @institution.id, name: @institution.name, image_key: @institution.image_key }
   end
 
   # POST institutions
@@ -32,20 +36,22 @@ class InstitutionsController < ApplicationController
 
   # PATCH/PUT institutions/1
   def update
+    return unless @institution
+
+    update_params = update_institution_params
+
     if params[:new_institution_image].present?
       s3_service = S3Service.new
-      uploaded_image_url = s3_service.upload_image(:institution_image, @institution.id, params[:new_institution_image])
+      uploaded_image_key = s3_service.upload_image(:institution_image, @institution.id, params[:new_institution_image])
 
-      old_image_key = @institution.image_url
+      old_image_key = @institution.image_key
+      s3_service.delete_image(:institution_image, @institution.id, old_image_key) if old_image_key.present?
 
-      s3_service.delete_image(:institution_image, @institution.id, old_image_key)
-
-      params[:image_url] = uploaded_image_url
-      params.delete(:new_institution_image)
+      update_params = update_params.merge(image_key: uploaded_image_key)
     end
 
-    if @institution.update(update_institution_params)
-      render json: @institution
+    if @institution.update(update_params)
+      render json: @institution, status: :ok
     else
       render json: @institution.errors, status: :unprocessable_entity
     end
@@ -60,17 +66,19 @@ class InstitutionsController < ApplicationController
   end
 
   private
-    def set_institution
-      @institution = Institution.find(params.expect(:id))
-    rescue ActiveRecord::RecordNotFound
+  def set_institution
+    @institution = Institution.find_by(id: params[:id])
+    unless @institution
       render json: { error: 'Institution not found' }, status: :not_found
     end
+  end
+
 
     def set_own_institution
-      profile_id = request.headers["X-Profile-ID"]
+      institution_id = request.headers["X-Institution-ID"]
 
-      if profile_id.present?
-        @institution = Institution.find(profile_id)
+      if institution_id.present?
+        @institution = Institution.find_by(id: institution_id)
       else
         render json: { error: "Profile ID header missing" }, status: :bad_request
       end
@@ -80,10 +88,10 @@ class InstitutionsController < ApplicationController
     end
 
     def create_institution_params
-      params.expect(institution: [ :name, :domain, :image_url ])
+      params.expect(institution: [ :name, :domain, :image_key ])
     end
 
     def update_institution_params
-      params.expect(institution: [ :name, :domain, :image_url, :new_institution_image ])
+      params.require(:institution).permit(:name, :domain, :uploaded_image )
     end
 end
